requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
defusedxml==0.6.0
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
googleads==26.0.0
idna==2.10
isodate==0.6.0
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.3
numpy==1.19.5
oauthlib==3.1.0
pandas==1.2.0
Pillow==8.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
PyYAML==5.3.1
pyzbar==0.1.8
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
xmltodict==0.12.0
zeep==4.0.0
